requirements.txt
requests>=2.23.0
PyYAML>=6.0
